+++ src/module/item/item/index.js
@@ -1,1 +1,2 @@
-export { TorchbearerItemSheet } from './sheet';
+export { TorchbearerItem } from "./item";
+export { TorchbearerItemSheet } from "./sheet";
