+++ src/module/item/spell/index.js
@@ -1,1 +1,2 @@
+export { TorchbearerSpell } from "./item";
 export { TorchbearerSpellSheet } from "./sheet";
