+++ src/module/item/spell/item.js
@@ -1,0 +1,3 @@
+import { TorchbearerBaseItem } from "../base-item";
+
+export class TorchbearerSpell extends TorchbearerBaseItem {}
